Fix command execution

+++ source/input.h
@@ -1,4 +1,4 @@
 #include <stdio.h>
 #include <sys/types.h>
 
-char* readLine(char* line, FILE *stream, uint maxLength);
+char *readLine(char *line, FILE *stream, ulong maxLength);
